[WCNCR00256074] misc: fix coverity 4601545
[Description]
Fix coverity 4601545

issue : CERT-C Characters and Strings

MTK-Commit-Id: aafe9ce6478a4376894a1fb0d8c4a80c8c9ac384

CR-Id: WCNCR00256074
Feature: misc
Reviewed-on: https://gerrit.mediatek.inc/c/neptune/wlan_driver/gen4m/+/5661243
CheckCourse: srv_pis <srv_pis@mediatek.com>
CheckCRInfo: srv_pis <srv_pis@mediatek.com>
CheckForbiddenKeyword: srv_pis <srv_pis@mediatek.com>
CheckPatch: srv_pis <srv_pis@mediatek.com>
Build: srv_neptune_adm <srv_neptune_adm@mediatek.com>
Test: srv_mspautosanity
(cherry picked from commit 952255c17ae0f1c56447f20c62a5d56fdd30e90b)
Reviewed-on: https://gerrit.mediatek.inc/c/neptune/wlan_driver/gen4m/+/5715956
Build: srv_preflight_a001 <srv_preflight_a001@mediatek.com>

+++ mgmt/saa_fsm.c
@@ -147,11 +147,11 @@ saaFsmSteps(IN struct ADAPTER *prAdapter,
 
 	do {
 		u4AuthAssocState = prStaRec->eAuthAssocState;
-		if (u4AuthAssocState < AA_STATE_NUM
-				&& eNextState < AA_STATE_NUM) {
+		if ((u4AuthAssocState < AA_STATE_NUM)
+				&& ((uint32_t) eNextState < AA_STATE_NUM)) {
 			DBGLOG(SAA, STATE, "[SAA]TRANSITION: [%s] -> [%s]\n",
 				apucDebugAAState[u4AuthAssocState],
-				apucDebugAAState[eNextState]);
+				apucDebugAAState[(uint32_t)eNextState]);
 		}
 
 		ePreviousState = prStaRec->eAuthAssocState;
